[Validator] Moved constraints Optional and Required to the Constraints\ namespace
