+++ src/Symfony/Component/Validator/Constraints/NotBlank.php
@@ -16,6 +16,8 @@ use Symfony\Component\Validator\Constraint;
 /**
  * @Annotation
  *
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ *
  * @api
  */
 class NotBlank extends Constraint
