+++ src/Symfony/Component/Validator/Constraints/NotBlankValidator.php
@@ -17,6 +17,8 @@ use Symfony\Component\Validator\ConstraintValidator;
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
  *
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ *
  * @api
  */
 class NotBlankValidator extends ConstraintValidator
