+++ src/Symfony/Component/Validator/Constraints/Optional.php
@@ -11,16 +11,11 @@
 
 namespace Symfony\Component\Validator\Constraints;
 
-use Symfony\Component\Validator\Constraint;
-
 /**
  * @Annotation
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
- *
- * @api
  */
-class DateTime extends Constraint
+class Optional extends Existence
 {
-    public $message = 'This value is not a valid datetime.';
 }
