+++ src/Symfony/Component/Validator/Constraints/Time.php
@@ -16,6 +16,8 @@ use Symfony\Component\Validator\Constraint;
 /**
  * @Annotation
  *
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ *
  * @api
  */
 class Time extends Constraint
